created a file for helper functions
